Final/requirements_phase1.txt
aiohttp==4.0.0a1
#allennlp==2.10.0
allennlp-models==2.10.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bleach==5.0.1
#blis==0.7.8
boto3==1.24.38
botocore==1.27.38
brotlipy==0.7.0
cached-path==1.1.5
cachetools==5.2.0
catalogue==1.0.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
conllu==4.4.2
cryptography==37.0.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.31
#datasets==2.4.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docker-pycreds==0.4.0
#en-core-web-sm==2.3.1
entrypoints==0.4
fairscale==0.4.6
fastjsonschema==2.16.1
filelock==3.7.1
fonttools==4.34.4
fsspec==2022.5.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
graphviz==0.20.1
#h5py==3.7.0
huggingface-hub==0.8.1
icecream==2.1.3
idna==3.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonnet==0.18.0
jsonschema==2.6.0
jupyter-client==7.2.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
langcodes==3.3.0
lmdb==1.3.0
MarkupSafe==2.1.1
#matplotlib==3.5.2
#matplotlib-inline==0.1.3
mistune==0.8.4
#mkl-fft==1.3.1
#mkl-random==1.2.2
#mkl-service==2.4.0
more-itertools==8.13.0
multidict==4.7.6
multiprocess==0.70.13
murmurhash==1.0.7
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.23.1
neuralcoref==4.0
nltk==3.7
notebook==6.4.12
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
plac==0.9.6
pluggy==1.0.0
preshed==3.0.6
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-rouge==1.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytextrank==3.2.4
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
regex==2022.7.25
requests==2.28.1
responses==0.18.0
rich==12.1.0
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.1.1
scipy==1.8.1
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.96
sentry-sdk==1.8.0
setproctitle==1.3.0
setuptools==61.2.0
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy #spacy==2.3.7
#spacy-legacy #spacy-legacy==3.0.9
#spacy-loggers==1.0.3
srsly==1.0.5
stack-data==0.2.0
tensorboardX==2.5.1
termcolor==1.1.0
terminado==0.15.0
thinc==7.4.5
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.10.0 #torch==1.12.0
torchaudio==0.12.0
#torchvision==0.13.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wandb==0.12.21
wasabi==0.10.0
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
widgetsnbextension==3.6.1
word2number==1.1
xxhash==3.0.0
yarl==1.7.2
